Thu, Nov 16, 2023, 1:48 PM +02:00

+++ meteorites/import.sql
@@ -16,11 +16,11 @@ CREATE TABLE meteorites_temp (
     "id" INT,
     "nametype" TEXT,
     "class" TEXT,
-    "mass" NUMERIC(10, 2),
+    "mass" NUMERIC,
     "discovery" TEXT,
     "year" INT,
-    "lat" NUMERIC(10, 2),
-    "long" NUMERIC(10, 2)
+    "lat" NUMERIC,
+    "long" NUMERIC
 );
